dspy-production-blueprint/requirements.txt
dspy-ai
openai
fastapi
uvicorn
python-dotenv
pydantic-settings
